-bug: Deleting large number of hosts results in SQL errors due to MySQL buffer overrun -bug: Re-Apply Suggested Data Source names results in a SQL error for rrd_maximum='query_ifSpeed'
git-svn-id: svn+ssh://repo.cacti.net/var/data/svnroot/cacti/cacti/branches/0.8.7@4912 860744bd-22fc-0310-8c96-e9fe5004b5ca

+++ lib/api_data_source.php
@@ -55,27 +55,49 @@ function api_data_source_remove_multi($local_data_ids) {
 			}
 
 			$i++;
-		}
 
-		$data_template_data_ids = db_fetch_assoc("SELECT id
-			FROM data_template_data
-			WHERE local_data_id IN ($ids_to_delete)");
+			if (($i % 1000) == 0) {
+				$data_template_data_ids = db_fetch_assoc("SELECT id
+					FROM data_template_data
+					WHERE local_data_id IN ($ids_to_delete)");
 
-		if (sizeof($data_template_data_ids)) {
-			foreach($data_template_data_ids as $data_template_data_id) {
-				if ($j == 0) {
-					$dtd_ids_to_delete .= $data_template_data_id["id"];
-				}else{
-					$dtd_ids_to_delete .= ", " . $data_template_data_id["id"];
+				if (sizeof($data_template_data_ids)) {
+					foreach($data_template_data_ids as $data_template_data_id) {
+						if ($j == 0) {
+							$dtd_ids_to_delete .= $data_template_data_id["id"];
+						}else{
+							$dtd_ids_to_delete .= ", " . $data_template_data_id["id"];
+						}
+
+						$j++;
+
+						if ($j % 1000) {
+							db_execute("DELETE FROM data_template_data_rra WHERE data_template_data_id IN ($dtd_ids_to_delete)");
+							db_execute("DELETE FROM data_input_data WHERE data_template_data_id IN ($dtd_ids_to_delete)");
+
+							$dtd_ids_to_delete = "";
+							$j = 0;
+						}
+					}
+
+					if ($j > 0) {
+						db_execute("DELETE FROM data_template_data_rra WHERE data_template_data_id IN ($dtd_ids_to_delete)");
+						db_execute("DELETE FROM data_input_data WHERE data_template_data_id IN ($dtd_ids_to_delete)");
+					}
 				}
 
-				$j++;
+				db_execute("DELETE FROM data_template_data WHERE local_data_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM data_template_rrd WHERE local_data_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM poller_item WHERE local_data_id IN ($ids_to_delete)");
+				db_execute("DELETE FROM data_local WHERE id IN ($ids_to_delete)");
+
+				$i = 0;
+				$ids_to_delete = "";
 			}
-
-			db_execute("DELETE FROM data_template_data_rra WHERE data_template_data_id IN ($dtd_ids_to_delete)");
-			db_execute("DELETE FROM data_input_data WHERE data_template_data_id IN ($dtd_ids_to_delete)");
 		}
+	}
 
+	if ($i > 0) {
 		db_execute("DELETE FROM data_template_data WHERE local_data_id IN ($ids_to_delete)");
 		db_execute("DELETE FROM data_template_rrd WHERE local_data_id IN ($ids_to_delete)");
 		db_execute("DELETE FROM poller_item WHERE local_data_id IN ($ids_to_delete)");
@@ -108,10 +130,20 @@ function api_data_source_disable_multi($local_data_ids) {
 			}
 
 			$i++;
+
+			if ($i % 1000) {
+				db_execute("DELETE FROM poller_item WHERE local_data_id IN ($ids_to_disable)");
+				db_execute("UPDATE data_template_data SET active='' WHERE local_data_id IN ($ids_to_disable)");
+
+				$i = 0;
+				$ids_to_delete = "";
+			}
 		}
 
-		db_execute("DELETE FROM poller_item WHERE local_data_id IN ($ids_to_disable)");
-		db_execute("UPDATE data_template_data SET active='' WHERE local_data_id IN ($ids_to_disable)");
+		if ($i > 0) {
+			db_execute("DELETE FROM poller_item WHERE local_data_id IN ($ids_to_disable)");
+			db_execute("UPDATE data_template_data SET active='' WHERE local_data_id IN ($ids_to_disable)");
+		}
 	}
 }
 
@@ -134,7 +166,7 @@ function api_reapply_suggested_data_source_title($local_data_id) {
 								$data_local["snmp_query_id"], $data_local["snmp_index"],
 								read_config_option("max_data_query_field_length"));
 				/* if there are no '|query' characters, all of the substitutions were successful */
-				if (!strstr($subs_string, "|query")) {
+				if ((!substr_count($subs_string, "|query")) && ($suggested_value["field_name"] == "name")) {
 					db_execute("update data_template_data set " . $suggested_value["field_name"] . "='" . $suggested_value["text"] . "' where local_data_id=" . $local_data_id);
 					/* once we find a working value, stop */
 					$suggested_values_data[$data_template_data_id]{$suggested_value["field_name"]} = true;
@@ -142,5 +174,5 @@ function api_reapply_suggested_data_source_title($local_data_id) {
 			}
 		}
 	}
-} 
+}
 ?>
